Set the batchSize of the oplog cursor to 20
This means more roundtrips to the server when slurping, but it also means
that less memory will be needed for the buffer.

This is a *hack* to avoid out-of-memory situations when many changes appear in
the oplog: each river reads the complete oplog, even though it discards most
of the entries.